# back/src/swagger.yaml
openapi: "3.0.0"
info:
  title: Photo Gallery API
  version: "1.0.0"
servers:
  - url: http://localhost:3000
    description: Local development
tags:
  - name: Auth
    description: 認証関連エンドポイント
  - name: Users
    description: ユーザー情報取得
  - name: Images
    description: 画像 CRUD / シェア
  - name: Likes
    description: いいね関連機能

paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Email＋PasswordでログインしてJWT取得
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: 認証トークン
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string

  /auth/refresh:
    post:
      tags: [Auth]
      summary: リフレッシュトークンでアクセストークン再発行
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: 新しい認証トークン
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string

  /users/me:
    get:
      tags: [Users]
      summary: ログイン中のユーザー情報取得
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User オブジェクト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  /images:
    get:
      tags: [Images]
      summary: 画像一覧取得（ページネーション）
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 画像リスト＋総件数
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Image'
                  total:
                    type: integer
    post:
      tags: [Images]
      summary: 画像アップロード（管理者のみ）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: 作成された Image オブジェクト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'

  /images/{id}:
    get:
      tags: [Images]
      summary: 画像詳細取得
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Image オブジェクト
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'

  /images/{id}/share:
    get:
      tags: [Images]
      summary: SNS向けシェア用URL取得
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: platform
          required: true
          schema:
            type: string
            enum: [twitter, facebook, line]
      responses:
        '200':
          description: シェア用 URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri

  /images/{id}/likes:
    post:
      tags: [Likes]
      summary: 画像にいいね
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: いいね成功
    delete:
      tags: [Likes]
      summary: 画像のいいね解除
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: いいね解除成功

  /users/me/likes:
    get:
      tags: [Likes]
      summary: ログインユーザーのいいね画像一覧取得
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Image オブジェクトの配列
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Image'

  /users/me/likes/download:
    get:
      tags: [Likes]
      summary: いいね済み画像一括ダウンロード用 ZIP URL 取得
      security:
        - bearerAuth: []
      responses:
        '200':
          description: ダウンロード用 URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri

  /users/me/likes/urls:
    get:
      tags: [Likes]
      summary: いいね済み画像の公開 URL 一覧取得
      security:
        - bearerAuth: []
      responses:
        '200':
          description: URL の文字列配列
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  format: uri

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Image:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        url:
          type: string
          format: uri
        storagePath:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time